update: Added support for floating-point calculations for discounts

+++ internal/logic/public/portal/tool.go
@@ -7,14 +7,14 @@ import (
 )
 
 func getDiscount(discounts []types.SubscribeDiscount, inputMonths int64) float64 {
-	var finalDiscount int64 = 100
+	var finalDiscount float64 = 100
 
 	for _, discount := range discounts {
 		if inputMonths >= discount.Quantity && discount.Discount < finalDiscount {
 			finalDiscount = discount.Discount
 		}
 	}
-	return float64(finalDiscount) / float64(100)
+	return finalDiscount / float64(100)
 }
 
 func calculateCoupon(amount int64, couponInfo *coupon.Coupon) int64 {
